+++ packages/ListBoxWithTags/src/ListBoxWithTags.js
@@ -33,9 +33,22 @@ export default function ListBoxWithTags(props) {
   const { t } = useI18n();
   const refDivRoot = React.useRef(null);
   const refFilter = React.useRef(null);
+  const refListBox = React.useRef(null);
   const [idListBoxContent] = React.useState(() => `list-box_${uuidv4()}__content`);
+  const [lastSelectedOption, setLastSelectedOption] = React.useState(null);
 
   const validSize = Object.values(ListBox.types.size).includes(size) ? size : ListBox.types.size.MEDIUM;
+
+  function handleLastSelectedOption(selectedOption, options) {
+    const lastIndex = Object.keys(options).length - 1;
+    if (lastIndex === 0 || refFilter.current.textSearch || options[lastIndex - 1].hasLabel === undefined) {
+      setLastSelectedOption(null);
+    } else if (selectedOption === lastIndex) {
+      setLastSelectedOption(lastIndex - 1);
+    } else {
+      setLastSelectedOption(selectedOption);
+    }
+  }
 
   function handleKeyDown(event) {
     const label = event.target.value;
@@ -47,13 +60,21 @@ export default function ListBoxWithTags(props) {
     ) {
       event.stopPropagation();
       onAddCustomOption(label);
-      refFilter.current.reset();
+      if (refFilter.current) {
+        refFilter.current.reset();
+        refFilter.current.focus();
+      }
     }
   }
 
   function handleChange(...args) {
+    const [newSelectedOptions, options] = args;
     if (selectedOptions !== null && "length" in selectedOptions) {
-      refFilter.current.reset();
+      handleLastSelectedOption(newSelectedOptions[0], options);
+      if (refFilter.current) {
+        refFilter.current.reset();
+        refFilter.current.focus();
+      }
     }
 
     onChange(...args);
@@ -79,11 +100,15 @@ export default function ListBoxWithTags(props) {
     "ListBox.Trigger": {},
     "ListBox.Box": {},
     "ListBox.Filter": {},
+    "ListBox.Content": {},
   };
 
   const filteredChildren = React.useMemo(() => {
     return React.Children.map(children, child => {
-      if (child && ["ListBox.Trigger", "ListBox.Box", "ListBox.Filter"].includes(child.type.displayName)) {
+      if (
+        child &&
+        ["ListBox.Trigger", "ListBox.Box", "ListBox.Filter", "ListBox.Content"].includes(child.type.displayName)
+      ) {
         const { children, ...moreProps } = child.props;
         extendedProps[child.type.displayName] = moreProps;
         return null;
@@ -92,9 +117,18 @@ export default function ListBoxWithTags(props) {
     });
   }, [children, extendedProps]);
 
+  React.useEffect(() => {
+    if (refListBox.current && refFilter.current) {
+      if (lastSelectedOption !== null && !refFilter.current.textSearch) {
+        refListBox.current.setFocusOptionByIndex(lastSelectedOption);
+      }
+    }
+  });
+
   return (
     <div ref={refDivRoot}>
-      <ListBox isMulti size={validSize} onChange={handleChange} {...moreProps}>
+      <ListBox ref={refListBox} isMulti size={validSize} onChange={handleChange} {...moreProps}>
+        <ListBox.Content {...extendedProps["ListBox.Content"]} />
         <ListBox.Trigger {...extendedProps["ListBox.Trigger"]}>
           {(...[, , , attributes]) => <TriggerWithTags {...triggerProps} {...attributes} />}
         </ListBox.Trigger>