+++ packages/OverflowMenu/src/OverflowMenu.styles.js
@@ -1,4 +1,5 @@
-import styled from "styled-components";
+import styled, { css } from "styled-components";
+import Popover from "@paprika/popover";
 import tokens from "@paprika/tokens";
 import stylers from "@paprika/stylers";
 
@@ -10,3 +11,10 @@ export const Content = styled.div`
     ${stylers.lineHeight(-1)}
   }
 `;
+
+export const PopoverCard = styled(Popover.Card)(
+  ({ maxHeight }) => css`
+    max-height: ${Number.isNaN(Number(maxHeight)) ? maxHeight : `${maxHeight}px`};
+    overflow: auto;
+  `
+);