+++ packages/Panel/stories/examples/FocusLockPanel.js
@@ -4,6 +4,7 @@ import Panel from "../../src";
 export default function FocusLockPanel() {
   return (
     <Panel isOpen>
+      <Panel.Overlay />
       <Panel.Header>Header</Panel.Header>
       <Panel.Content>
         <input type="text" data-autofocus />