+++ packages/Pill/stories/Pill.stories.helpers.js
@@ -1,5 +1,0 @@
-export const storyStyles = `
-  > div {
-    margin-right: 8px;
-  }
-`;