+++ packages/ActionBar/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## 3.1.5
+
+### Patch Changes
+
+- Updated dependencies [009352de9]
+  - @paprika/icon@1.2.0
+  - @paprika/button@1.1.1
+  - @paprika/input@4.0.4
+  - @paprika/select@2.0.1
+  - @paprika/sortable@1.1.12
+
+## 3.1.5-next.0
+
+### Patch Changes
+
+- Updated dependencies [009352d]
+  - @paprika/icon@1.2.0-next.0
+  - @paprika/button@1.1.1-next.0
+  - @paprika/input@4.0.4-next.0
+  - @paprika/select@2.0.1-next.0
+  - @paprika/sortable@1.1.12-next.0
+
+## 3.1.4
+
+### Patch Changes
+
+- Updated dependencies [ae0d6a0]
+  - @paprika/l10n@1.1.11
+  - @paprika/input@4.0.3
+  - @paprika/sortable@1.1.11
+
 ## 3.1.4-next.0
 
 ### Patch Changes