+++ packages/Confirmation/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 1.1.6
+
+### Patch Changes
+
+- @paprika/button@1.1.1
+
+## 1.1.6-next.0
+
+### Patch Changes
+
+- @paprika/button@1.1.1-next.0
+
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [ae0d6a0]
+  - @paprika/l10n@1.1.11
+
 ## 1.1.5-next.0
 
 ### Patch Changes