+++ packages/DataField/CHANGELOG.md
@@ -1,5 +1,32 @@
 <!-- content -->
 
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [659ebe184]
+- Updated dependencies [009352de9]
+  - @paprika/data-grid@1.0.14
+  - @paprika/icon@1.2.0
+  - @paprika/button@1.1.1
+
+## 1.0.18-next.0
+
+### Patch Changes
+
+- Updated dependencies [009352d]
+  - @paprika/icon@1.2.0-next.0
+  - @paprika/button@1.1.1-next.0
+  - @paprika/data-grid@1.0.14-next.0
+
+## 1.0.17
+
+### Patch Changes
+
+- Updated dependencies [ae0d6a0]
+  - @paprika/l10n@1.1.11
+  - @paprika/data-grid@1.0.13
+
 ## 1.0.17-next.0
 
 ### Patch Changes