+++ packages/Modal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.0.12
+
+### Patch Changes
+
+- @paprika/button@1.1.1
+
+## 1.0.12-next.0
+
+### Patch Changes
+
+- @paprika/button@1.1.1-next.0
+
 ## 1.0.11
 
 ### Patch Changes