+++ packages/Toast/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+## 1.1.6
+
+### Patch Changes
+
+- bacebc9: Refactoring internal Toast states and fixing canAutoClose issue for controlled components
+
+## 1.1.6-next.0
+
+### Patch Changes
+
+- bacebc9: Refactoring internal Toast states and fixing canAutoClose issue for controlled components
+
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [009352de9]
+  - @paprika/icon@1.2.0
+  - @paprika/button@1.1.1
+
+## 1.1.5-next.0
+
+### Patch Changes
+
+- Updated dependencies [009352d]
+  - @paprika/icon@1.2.0-next.0
+  - @paprika/button@1.1.1-next.0
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [ae0d6a0]
+  - @paprika/l10n@1.1.11
+
 ## 1.1.4-next.0
 
 ### Patch Changes